test(1): fix find headSha call
The gh pr view call had `message-correlation` in it, this was a leftover
from trying out this call locally. It refers to a specific branch. Since
a pr was already created from the current checked out branch, we can
simply remove any specification of a branch and let gh determine which
pr we mean.