test(2): should create pr from fork
After adding the remote, we should be able to push a commit/branch to
the fork repo. Once that's done, we should be able to create a PR from
the fork's new branch to the origin (upstream)'s main branch.

Note that upstream is just called origin, because that is already
checked out when this test case is run.