fix: delete revert branch with merge
If the revert PR is not merged, but the branch is deleted afterwards,
then the PR is closed. This makes it less visible that cleanup has
failed.

We can fix this by not deleting afterwards, but during the merge. If the
merge fails than we still have the PR open.